pyproject.toml
# =============================================================================
# Ref: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
#
# Os pontos de atenção estão comentados próximo das chaves, aqui só um lembrete
#
# Atenção: As versões do Python devem ser consistentes entre:
#   - project.requires-python
#   - tool.ruff.target-version
#   - tool.pyright.pythonVersion
#
# Atenção: O nome do pacote principal deve ser consistente entre:
#   - project.scripts
#   - tool.ruff.lint.isort.known-first-party
#   - A pasta dentro de 'src/'
# =============================================================================
[project]
name = "uv-model"
version = "0.1.0"
description = "Add description here"
readme = "README.md"
# license = "MIT"
authors = [{ name = "Jadson Nascimento", email = "jadson.nascimento060@academico.ifs.edu.br" }]
requires-python = ">=3.12, <4.0"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=1.0.16",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.28",
    "openai>=1.99.1",
    "pandas>=2.3.1",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.1",
]

[project.urls]
# Homepage = "https://www.example.com.br/"
# Documentation = "https://example.org"
# Repository = "https://github.com/me/spam.git"
# Issues = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pyright>= 1.1.403",
    "pytest>=8.4.1",
    "pytest-xdist >= 3.8.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
    "taskipy>=1.14.1",
]

[project.scripts]
# Define os comandos de console. A chave ('src') é o comando.
# O valor aponta para 'nome_do_pacote.nome_do_modulo:nome_da_funcao'.
# Ex: 'package.main:run' -> src/package/main.py (e a função run lá dentro).
# Lembre-se de sincronizar o nome do pacote com 'known-first-party' do Ruff.
c = "package.module:script"
b = "src.app:main"
a = "main:main"

# =========================================================================== #
#                          Lint e formatação (Ruff)                           #
# =========================================================================== #
[tool.ruff]
line-length = 79
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]

[tool.ruff.lint]
preview = true
select = [
    "ASYNC",
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "I",
    "ICN",
    "ISC",
    'PL',
    'PT',
    "Q",
    "RET",
    "RSE",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TC",
    "TD",
    "TRY",
    "UP",
    "W",
    "YTT",
    "RUF",
    "N",
]
ignore = ["T201", "COM812"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
line-ending = "lf"
preview = true

[tool.ruff.lint.isort]
# Ensina ao Ruff qual é o pacote principal do seu projeto.
# Deve ser o mesmo nome da pasta dentro de 'src/'.
known-first-party = ["src"]

# =========================================================================== #
#                           Tipagem (Pyright)                                 #
# =========================================================================== #
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13.5"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

# =========================================================================== #
#                              Testes (Pytest)                                #
# =========================================================================== #
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
# Adiciona 'src' ao path do Python durante os testes. Isso permite que
# os testes em `tests/` importem os módulos de `src/` sem erros.
pythonpath = ["src"]
testpaths = ["tests"]

# =========================================================================== #
#                              Tasks (TaskiPy)                                #
# =========================================================================== #
[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'uv run app.py'
runs = 'uv run main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=poc_mcp -vv'
post_test = 'coverage html'
teste = 'pytest -v -s --color=yes --tb=short'

# =========================================================================== #
#                            Build e Setuptools                               #
# =========================================================================== #
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.7",
    "taskipy>=1.14.1",
    "uv-model",
]

[tool.setuptools.packages.find]
# Informa ao setuptools para encontrar os pacotes dentro da pasta 'src'.
# Essencial para a estrutura de layout com 'src/'.
where = ["src"]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
uv-model = { workspace = true }

[tool.setuptools]
# Mapeia o diretório raiz de pacotes para 'src'.
package-dir = { "" = "src" }
